Add storybook to the project and update dependencies.
Updated the package.json and package-lock.json files to include Storybook packages and scripts to run Storybook. This is done to facilitate a more interactive UI development environment as Storybook enables building and testing of UI components in isolation. Also, updated the versions of some packages and added specific versions to others to ensure compatibility and stability.